requirements.txt
aiodns==3.2.0
aiosmtpd==1.4.6
asgiref==3.8.1
atpublic==5.0
attrs==24.2.0
backports.tarfile==1.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
cssselect==1.2.0
cssutils==2.11.1
dkimpy==1.1.8
dnspython==2.6.1
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Executor==1.0.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.7
importlib_metadata==8.2.0
iniconfig==2.0.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.2
jeepney==0.8.0
Jinja2==3.1.4
keyring==25.3.0
lxml==5.3.0
MarkupSafe==2.1.5
more-itertools==10.4.0
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
premailer==3.10.0
pycares==4.4.0
pycodestyle==2.12.1
pycparser==2.22
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
SecretStorage==3.3.3
SQLAlchemy==2.0.32
SQLAlchemy-Utils==0.41.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.2
verify-email==2.4.3
Werkzeug==3.0.3
WTForms==3.1.2
yagmail==0.15.293
zipp==3.20.0
